requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
colorama==0.4.6
distro==1.9.0
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.58.1
pillow==11.1.0
pydantic==2.10.4
pydantic_core==2.27.2
PyPDF2==3.0.1
python-telegram-bot==21.9
rfc3986==1.5.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
